revert(security): drop response-file and feedback-gateway fixes tracked elsewhere
Two findings from this review already have open PRs with better fixes, so
this branch stops duplicating them.

ENG-1981 (#8647) — cross-tenant response file deletion. That PR fixes the
source: all four management response routes validate uploads with
validateClientFileUploads, and the weak extension-only validateFileUploads
is deleted, so a foreign storage URL never persists in the first place.
This branch only guarded the deletion sink, which still let the bad URL be
stored and rendered in views and exports.

ENG-1980 (#8648) — API-key organizationAccess bypassing workspace scoping
on feedback records. That PR removes the organizationAccess shortcuts
outright, which is the correct model: org-level access control governs org
management, not workspace data. This branch only bound the shortcut to the
directory's organization, leaving a same-org hole where an accessControl
key with no workspace permission still reached the records.

The manage-for-deletes change in feedback-records-gateway.ts stays; it is a
separate finding.
