+++ apps/web/modules/api/v2/management/responses/[responseId]/lib/utils.ts
@@ -1,8 +1,10 @@
 import { Response, Survey } from "@formbricks/database/prisma";
+import { logger } from "@formbricks/logger";
 import { Result, okVoid } from "@formbricks/types/error-handlers";
 import { TSurveyQuestionTypeEnum } from "@formbricks/types/surveys/types";
 import { ApiErrorResponseV2 } from "@/modules/api/v2/types/api-error";
-import { deleteResponseFileUrls } from "@/modules/storage/lib/delete-response-files";
+import { deleteFile } from "@/modules/storage/service";
+import { parseStorageFileUrl } from "@/modules/storage/utils";
 
 export const findAndDeleteUploadedFilesInResponse = async (
   responseData: Response["data"],
@@ -21,7 +23,20 @@ export const findAndDeleteUploadedFilesInResponse = async (
     .filter(([questionId]) => fileUploadQuestions.has(questionId))
     .flatMap(([, questionResponse]) => questionResponse as string[]);
 
-  await deleteResponseFileUrls(fileUrls, workspaceId);
+  const deletionPromises = fileUrls.map(async (fileUrl) => {
+    try {
+      const storageFile = parseStorageFileUrl(fileUrl);
+
+      if (!storageFile) {
+        throw new Error(`Invalid storage file URL: ${fileUrl}`);
+      }
+      return deleteFile(storageFile.storageId, storageFile.accessType, storageFile.fileName, workspaceId);
+    } catch (error) {
+      logger.error({ error, fileUrl }, "Failed to delete file");
+    }
+  });
+
+  await Promise.all(deletionPromises);
 
   return okVoid();
 };
